// src/app/shared/models/sum-out/summary-trading.model.ts
export interface SummaryTrading {
    totals: {
      buyBg: number;
      buyKg: number;
      buyOz: number;
      buyPl: number;
      buyOuts: number;
      buyAmount: number;
      sellBg: number;
      sellKg: number;
      sellOz: number;
      sellPl: number;
      sellOuts: number;
      sellAmount: number;
      totalBg: number;
      totalKg: number;
      totalOz: number;
      totalPl: number;
      totalOuts: number;
      totalAmount: number;
    };
    summary: {
      buyBg: number;
      buyBgQty: number;
      buyKg: number;
      buyKgQty: number;
      buyOz: number;
      buyOzQty: number;
      buyUsd: number;
      totalBuyKg: number;
      totalBuyQty: number;
      totalBuyIMEXKg: number;
      totalBuyIMEXQty: number;
  
      sellBg: number;
      sellBgQty: number;
      sellKg: number;
      sellKgQty: number;
      sellOz: number;
      sellOzQty: number;
      sellUsd: number;
      totalSellKg?: number;
      totalSellQty?: number;
      totalSellIMEXKg?: number;
      totalSellIMEXQty?: number;
    };
  }
  