// src/app/shared/models/sum-out/summarySettle.model.ts
export interface SummarySettle {
    buyBg: number;
    buyBgQty: number;
    buyKg: number;
    buyKgQty: number;
  
    buyUsd: number;
    totalBuyKg: number;
    totalBuyQty: number;
  
    sellBg: number;
    sellBgQty: number;
    sellKg: number;
    sellKgQty: number;
  
    sellUsd: number;
    totalSellKg: number;
    totalSellQty: number;
  }